refactor(fwa): extract mailConfig helpers into dedicated module
- move mail config types/parsing and input-json conversion from Fwa.ts to src/commands/fwa/mailConfig.ts
- move mail reference helpers (primary ref, target collection, recovery rewrite) into reusable fwa/mailConfig module
- move force message-type parsing and discord message URL helpers into reusable module
- keep runtime behavior unchanged while reducing Fwa.ts surface area and improving reuse
